+++ src/WorkspaceContext.ts
@@ -108,7 +108,7 @@ export class WorkspaceContext implements ExternalWorkspaceContext, Disposable {
         this.testRunManager = new TestRunManager();
         this.loggerFactory = new SwiftLoggerFactory(extensionContext.logUri);
         this.statusItem = new StatusItem();
-        this.buildStatus = new SwiftBuildStatus();
+        this.buildStatus = new SwiftBuildStatus(this.statusItem);
         this.languageClientManager = new LanguageClientToolchainCoordinator(this, {
             onDocumentSymbols: (folder, document, symbols) => {
                 folder.onDocumentSymbols(document, symbols);
@@ -251,6 +251,7 @@ export class WorkspaceContext implements ExternalWorkspaceContext, Disposable {
             this.documentation,
             this.languageClientManager,
             this.logger,
+            this.statusItem,
             this.buildStatus,
             this.projectPanel,
         ];